Give clockwork a dedicated provider
