+++ src/class-clockwork-provider.php
@@ -1,0 +1,104 @@
+<?php
+
+namespace Clockwork_For_Wp;
+
+use Clockwork\Authentication\AuthenticatorInterface;
+use Clockwork\Authentication\NullAuthenticator;
+use Clockwork\Authentication\SimpleAuthenticator;
+use Clockwork\Clockwork;
+use Clockwork\Storage\FileStorage;
+use Clockwork\Storage\SqlStorage;
+use Clockwork\Storage\StorageInterface;
+use Clockwork_For_Wp\Config;
+use Clockwork_For_Wp\Data_Source\Php;
+
+class Clockwork_Provider extends Base_Provider {
+	public function boot() {
+		// Ensures Clockwork is instantiated and all data sources are added on 'plugins_loaded'.
+		$this->plugin[ Clockwork::class ]; // @todo Move into conditional?
+
+		if ( $this->plugin->is_collecting_data() ) {
+			parent::boot();
+		}
+	}
+
+	public function register() {
+		$this->plugin[ Clockwork_Subscriber::class ] = function() {
+			return new Clockwork_Subscriber( $this->plugin );
+		};
+
+		$this->plugin[ Clockwork::class ] = function() {
+			$clockwork = new Clockwork();
+
+			$clockwork->addDataSource( $this->plugin[ Php::class ] );
+
+			foreach ( $this->plugin->get_enabled_data_sources() as $data_source ) {
+				$clockwork->addDataSource( $this->plugin[ $data_source['data_source_class'] ] );
+			}
+
+			$clockwork->setAuthenticator( $this->plugin[ AuthenticatorInterface::class ] );
+			$clockwork->setStorage( $this->plugin[ StorageInterface::class ] );
+
+			return $clockwork;
+		};
+
+		$this->plugin[ AuthenticatorInterface::class ] = function() {
+			$config = $this->plugin[ Config::class ]->get( 'authentication' );
+
+			if ( ! $config['enabled'] ?? false ) {
+				return new NullAuthenticator();
+			}
+
+			$driver = $config['driver'] ?? 'simple';
+			$factory_id = $config['drivers'][ $driver ]['class'] ?? SimpleAuthenticator::class;
+
+			return $this->plugin[ $factory_id ]( $config['drivers'][ $driver ]['config'] ?? [] );
+		};
+
+		$this->plugin[ SimpleAuthenticator::class ] = $this->plugin->protect(
+			function( array $config ) {
+				if ( ! array_key_exists( 'password', $config ) ) {
+					throw new \InvalidArgumentException( '@todo' );
+				}
+
+				return new SimpleAuthenticator( $config['password'] );
+			}
+		);
+
+		$this->plugin[ StorageInterface::class ] = function() {
+			$config = $this->plugin[ Config::class ]->get( 'storage' );
+			$driver = $config['driver'] ?? 'file';
+			$factory_id = $config['drivers'][ $driver ]['class'] ?? FileStorage::class;
+
+			return $this->plugin[ $factory_id ]( $config['drivers'][ $driver ]['config'] ?? [] );
+		};
+
+		$this->plugin[ FileStorage::class ] = $this->plugin->protect( function( array $config ) {
+			if ( ! array_key_exists( 'path', $config ) ) {
+				throw new \InvalidArgumentException( '@todo' );
+			}
+
+			$dir_permissions = $config['dir_permissions'] ?? 0700;
+			$expiration = $config['expiration'] ?? null;
+
+			return new FileStorage( $config['path'], $dir_permissions, $expiration );
+		} );
+
+		$this->plugin[ SqlStorage::class ] = $this->plugin->protect( function( array $config ) {
+			if ( ! array_key_exists( 'dsn', $config ) ) {
+				throw new \InvalidArgumentException( '@todo' );
+			}
+
+			$table = $config['table'] ?? 'clockwork';
+			$username = $config['username'] ?? null;
+			$password = $config['password'] ?? null;
+			$expiration = $config['expiration'] ?? null;
+
+			return new SqlStorage( $config['dsn'], $table, $username, $password, $expiration );
+		} );
+	}
+
+	protected function subscribers() : array {
+		return [ Clockwork_Subscriber::class ];
+	}
+}
