+++ src/class-clockwork-subscriber.php
@@ -5,9 +5,8 @@ namespace Clockwork_For_Wp;
 use Clockwork\Clockwork;
 use Clockwork_For_Wp\Event_Management\Event_Manager;
 use Clockwork_For_Wp\Event_Management\Managed_Subscriber;
-use Clockwork_For_Wp\Plugin;
 
-class Plugin_Subscriber implements Managed_Subscriber {
+class Clockwork_Subscriber implements Managed_Subscriber {
 	protected $plugin;
 
 	public function __construct( Plugin $plugin ) {
